Add automatically updating version to dataset
Adds a version.txt file for tracking the semantic version of the data. Semantic
versioning for data follows https://frictionlessdata.io/specs/patterns/

By default anytime the repository is updated the minor version will be
incremented since this is the most common use case. Major version bumps, for
changes in data structure, can be indicated by including `[major]` in the commit
message. Patch version bumps, for fixes to existing data, can be indicated by
including `[minor]` in the commit message. If no changes to the data are made
include `[no version bump]` in the commit message.