Check correct commit for version bump tags when merging a PR [no version bump]
When a PR is merged the last commit is the merge commit. To check for version
bump tags we need to check the commit that was merged, not the merge commit.
This checks to see if the last commit was a PR merge and if so checks the next
commit back.