Add rodent census 541 data [minor] (#590)
* Add rodent census 541 data [minor]

* Let new moon script use predetermined new moon dates [minor]
The new moon script fetches official new moon dates using the lunar package. It adds them as they occur to the moon_dates.csv. Then, when the rodent data are added, it fetches those dates again to pair the new period with the closest unassigned new moon. Which shouldn't be a problem, they should be exactly the same. But I think, especially once the new moon is far in the past, it may arbitrarily choose one day or another when it spans both. So I've just switched that section of the script to use the new moon date from the moon_dates.csv, instead of recalculating it when it already exists.

---------

Co-authored-by: Glenda Yenni <glenda@weecology.org>